day11

+++ src/day11/solutions.h
@@ -4,4 +4,4 @@
 #include <vector>
 
 int PartOne(std::vector<std::string> input);
-long PartTwo(std::vector<std::string> input, const int weight);
+long long PartTwo(std::vector<std::string> input, const int weight);
